Rafactoring/Bug fixing
Fixed a bug in the Movement Indicator Spawn script causing the Indicator to spawn in the middle of a clicked object rather than on the ground.

Updated "PlayerMovement" Script to now just be "Player".
Made "Player" a Singleton.
Added Player Stat variables for use later.
Added Player State variables and functions as we were duplicating code in the now redundant PlayerState script.
Removed Redundant PlayerState script
Removed DoDamge script as not being used.